requirements.txt
spacy
langextract
requests
python-dotenv
langchain-openai